+++ kotlinbackend/src/main/kotlin/ViewKweet.kt
@@ -19,7 +19,7 @@ fun Route.viewKweet(dao: ViveDao, hashFunction: (String) -> String) {
         val date = System.currentTimeMillis()
         val code = if (user != null) call.securityCode(date, user, hashFunction) else null
 
-        val id = it.id.replace(",","").toInt()
+        val id = it.id.replace(",","").toLong()
         val kweet = dao.getKweet(id)
         call.respond(FreeMarkerContent("view-kweet.ftl", mapOf("user" to user, "kweet" to kweet, "date" to date, "code" to code), user?.userId ?: ""))
     }
